Fixed FEN Board Generation Last night I discovered that there was an additional problem with the board that only checked if there was a number then left 1 cell blank then proceeded to the next character in the string. It is fixed and I have two visual checks with the button "DOES STUFF".
I haven't verified that the algorithm for generating and interpreting
the string is "full-proof" as I pretty much just applied duct tape to
every problem I found after the initial write up.

However it seems like a piece can be moved and the board can be
generated. I will clean this up so we can start on the chess game mode.